Phase-3/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
pydantic==2.6.4
python-multipart==0.0.9
better-exceptions==0.3.2
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1